remove long long type

+++ kernel/Lyra2REv2.cl
@@ -48,8 +48,8 @@ typedef int sph_s32;
 typedef unsigned long sph_u64;
 typedef long  sph_s64;
 #else
-typedef unsigned long long sph_u64;
-typedef long long sph_s64;
+typedef unsigned long sph_u64;
+typedef long sph_s64;
 #endif
